+++ src/main/java/net/silentchaos512/lib/data/client/LibModelProvider.java
@@ -17,7 +17,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.data.CachedOutput;
 import net.minecraft.data.DataProvider;
 import net.minecraft.data.PackOutput;
-import net.minecraft.resources.ResourceLocation;
+import net.minecraft.resources.Identifier;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.level.block.Block;
@@ -53,10 +53,10 @@ public abstract class LibModelProvider implements DataProvider {
     }
 
     @Nullable
-    abstract protected BlockModelGenerators createBlockModelGenerators(Consumer<BlockModelDefinitionGenerator> blockStateOutput, ItemModelOutput itemModelOutput, BiConsumer<ResourceLocation, ModelInstance> modelOutput);
+    abstract protected BlockModelGenerators createBlockModelGenerators(Consumer<BlockModelDefinitionGenerator> blockStateOutput, ItemModelOutput itemModelOutput, BiConsumer<Identifier, ModelInstance> modelOutput);
 
     @Nullable
-    abstract protected ItemModelGenerators createItemModelGenerators(ItemModelOutput itemModelOutput, BiConsumer<ResourceLocation, ModelInstance> modelOutput);
+    abstract protected ItemModelGenerators createItemModelGenerators(ItemModelOutput itemModelOutput, BiConsumer<Identifier, ModelInstance> modelOutput);
 
     @Override
     public CompletableFuture<?> run(CachedOutput output) {
@@ -100,7 +100,8 @@ public abstract class LibModelProvider implements DataProvider {
 
         public CompletableFuture<?> save(CachedOutput p_388014_, PackOutput.PathProvider p_388192_) {
             Map<Block, BlockModelDefinition> map = Maps.transformValues(this.generators, BlockModelDefinitionGenerator::create);
-            Function<Block, Path> function = p_387598_ -> p_388192_.json(p_387598_.builtInRegistryHolder().key().location());
+            //noinspection deprecation
+            Function<Block, Path> function = p_387598_ -> p_388192_.json(p_387598_.builtInRegistryHolder().key().identifier());
             return DataProvider.saveAll(p_388014_, BlockModelDefinition.CODEC, function, map);
         }
     }
@@ -112,6 +113,11 @@ public abstract class LibModelProvider implements DataProvider {
         @Override
         public void accept(Item item, ItemModel.Unbaked model) {
             this.register(item, new ClientItem(model, ClientItem.Properties.DEFAULT));
+        }
+
+        @Override
+        public void accept(Item item, ItemModel.Unbaked model, ClientItem.Properties properties) {
+            this.register(item, new ClientItem(model, properties));
         }
 
         public void register(Item p_388205_, ClientItem p_388233_) {
@@ -131,7 +137,7 @@ public abstract class LibModelProvider implements DataProvider {
                 BuiltInRegistries.ITEM.listElements().filter(LibModelProvider.this::isModded).forEach(p_388426_ -> {
                     if (!this.copies.containsKey(p_388426_)) {
                         if (p_388426_.value() instanceof BlockItem blockitem && !this.itemInfos.containsKey(blockitem)) {
-                            ResourceLocation resourcelocation = ModelLocationUtils.getModelLocation(blockitem.getBlock());
+                            Identifier resourcelocation = ModelLocationUtils.getModelLocation(blockitem.getBlock());
                             this.accept(blockitem, ItemModelUtils.plainModel(resourcelocation));
                         }
                     }
@@ -148,16 +154,17 @@ public abstract class LibModelProvider implements DataProvider {
         }
 
         public CompletableFuture<?> save(CachedOutput output, PackOutput.PathProvider pathProvider) {
+            //noinspection deprecation
             return DataProvider.saveAll(
-                    output, ClientItem.CODEC, p_388594_ -> pathProvider.json(p_388594_.builtInRegistryHolder().key().location()), this.itemInfos
+                    output, ClientItem.CODEC, p_388594_ -> pathProvider.json(p_388594_.builtInRegistryHolder().key().identifier()), this.itemInfos
             );
         }
     }
 
-    static class SimpleModelCollector implements BiConsumer<ResourceLocation, ModelInstance> {
-        private final Map<ResourceLocation, ModelInstance> models = new HashMap<>();
+    static class SimpleModelCollector implements BiConsumer<Identifier, ModelInstance> {
+        private final Map<Identifier, ModelInstance> models = new HashMap<>();
 
-        public void accept(ResourceLocation location, ModelInstance instance) {
+        public void accept(Identifier location, ModelInstance instance) {
             Supplier<JsonElement> supplier = this.models.put(location, instance);
             if (supplier != null) {
                 throw new IllegalStateException("Duplicate model definition for " + location);
@@ -170,6 +177,6 @@ public abstract class LibModelProvider implements DataProvider {
     }
 
     private boolean isModded(Holder.Reference<?> reference) {
-        return reference.key().location().getNamespace().equals(this.modId);
+        return reference.key().identifier().getNamespace().equals(this.modId);
     }
 }