+++ src/main/java/net/silentchaos512/lib/data/recipe/ExtendedShapedRecipeBuilder.java
@@ -2,14 +2,14 @@ package net.silentchaos512.lib.data.recipe;
 
 import com.mojang.datafixers.util.Function5;
 import net.minecraft.advancements.*;
-import net.minecraft.advancements.critereon.RecipeUnlockedTrigger;
+import net.minecraft.advancements.criterion.RecipeUnlockedTrigger;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.recipes.RecipeBuilder;
 import net.minecraft.data.recipes.RecipeCategory;
 import net.minecraft.data.recipes.RecipeOutput;
+import net.minecraft.resources.Identifier;
 import net.minecraft.resources.ResourceKey;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
@@ -112,7 +112,7 @@ public abstract class ExtendedShapedRecipeBuilder<R extends CraftingRecipe> impl
 
         R recipe = createRecipe(id);
         AdvancementHolder advancementHolder = advancementBuilder != null
-                ? advancementBuilder.build(id.location().withPrefix("recipes/" + this.category.getFolderName() + "/"))
+                ? advancementBuilder.build(id.identifier().withPrefix("recipes/" + this.category.getFolderName() + "/"))
                 : null;
         output.accept(id, recipe, advancementHolder);
     }
@@ -137,7 +137,7 @@ public abstract class ExtendedShapedRecipeBuilder<R extends CraftingRecipe> impl
         return showNotification;
     }
 
-    public static ShapedRecipe vanillaFactory(ResourceLocation id, ExtendedShapedRecipeBuilder<ShapedRecipe> builder) {
+    public static ShapedRecipe vanillaFactory(Identifier id, ExtendedShapedRecipeBuilder<ShapedRecipe> builder) {
         // Basically the same as ShapedRecipeBuilder, but doesn't fail if advancement is missing
         return new ShapedRecipe(
                 Objects.requireNonNullElse(builder.group, ""),